+++ GuaranteedRate.Sextant/EncompassUtils/LoanDataUtils.cs
@@ -105,11 +105,30 @@ namespace GuaranteedRate.Sextant.EncompassUtils
                 AddLoanData(loanData, "milestones", ExtractMilestones(loan));
                 AddLoanData(loanData, "lastmodified", loan.LastModified.ToString(_DateFormat));
                 AddLoanData(loanData, "MachineUser", MachineUser.GetMachineUserIdentification());
-
+                AddLoanData(loanData, "attachments", ExtractLoanAttachments(loan));
+                
                 //restore state
                 loan.BorrowerPairs.Current = originalPair;
             }
             return loanData;
+        }
+
+
+        private static List<Dictionary<string, object>> ExtractLoanAttachments(Loan loan)
+        {
+            var docs = new List<Dictionary<string, object>>();
+            for (int i = 0; i < loan.Attachments.Count; i++)
+            {
+                var doc = new Dictionary<string, object>();
+                doc.Add("name", loan.Attachments[i].Name);
+                doc.Add("title", loan.Attachments[i].Title);
+                doc.Add("size", loan.Attachments[i].Size);
+                doc.Add("date", loan.Attachments[i].Date);
+                doc.Add("active", loan.Attachments[i].IsActive);
+
+                docs.Add(doc);
+            }
+            return docs;
         }
 
         private static void AddLoanData(IDictionary<string, object> loanData, string key, object value)
@@ -213,7 +232,6 @@ namespace GuaranteedRate.Sextant.EncompassUtils
                 fieldValues["LoanName"] = loan.LoanName;
                 fieldValues["LoanOfficerID"] = loan.LoanOfficerID;
                 fieldValues["LoanProcessorID"] = loan.LoanProcessorID;
-
                 Session session = loan.Session;
                 if (session != null)
                 {
@@ -224,6 +242,7 @@ namespace GuaranteedRate.Sextant.EncompassUtils
                         //just the smartClientId
                     fieldValues["SessionUserId"] = session.UserID;
                 }
+                
             }
             catch (Exception ex)
             {